requirements.txt
torch
opencv-python
Pillow
ultralytics
gradio
ipython